iOS: Fix a few instances of UB (#2428)
* Fix iOS 32-bit

* Fix a few invalid message sends on iOS

+++ src/platform_impl/ios/view.rs
@@ -262,7 +262,7 @@ unsafe fn get_view_class(root_view_class: &'static Class) -> &'static Class {
                         let scale_factor: CGFloat = msg_send![object, contentScaleFactor];
                         PhysicalPosition::from_logical::<(f64, f64), f64>(
                             (logical_location.x as _, logical_location.y as _),
-                            scale_factor,
+                            scale_factor as f64,
                         )
                     };
                     touch_events.push(EventWrapper::StaticEvent(Event::WindowEvent {
